FontBakeryCheckLogger class now in its own python script
(issue #1204)
